+++ mobile/lib/uis/core/buttons/big_button.dart
@@ -4,12 +4,16 @@ class BigButton extends StatelessWidget {
   final String label;
   final VoidCallback onPressed;
   final bool enabled;
+  final IconData? rightIcon;
+  final IconData? leftIcon;
 
   const BigButton({
     super.key,
     required this.label,
     required this.onPressed,
     required this.enabled,
+    this.rightIcon,
+    this.leftIcon,
   });
 
   @override
@@ -29,7 +33,20 @@ class BigButton extends StatelessWidget {
       style: style,
       child: Padding(
         padding: EdgeInsets.symmetric(vertical: 12),
-        child: Text(label, style: textStyle),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            if (leftIcon != null) ...[
+              Icon(leftIcon, size: 24),
+              SizedBox(width: 8),
+            ],
+            Text(label, style: textStyle),
+            if (rightIcon != null) ...[
+              SizedBox(width: 8),
+              Icon(rightIcon, size: 24),
+            ],
+          ],
+        ),
       ),
     );
   }