+++ mobile/lib/uis/core/messages/scaffold_snackbar_message.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+void showScaffoldSnackBarMessage(
+  BuildContext context, {
+  required String message,
+  bool isError = false,
+}) {
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(
+      content: Text(message),
+      backgroundColor: isError ? Colors.red : null,
+    ),
+  );
+}